Refactoring and Optimization
Refactoring is a formal and mechanical process, used to modify existing code in such a way that it does indeed become 'better' while
preserving the program's intended functionality. In addition to improving a program's overall design, the refactoring process tends to yield
code which is far easier to maintain and extend in the long run.